Clearing some commented prints.
