+++ src/terminal/highlight.c
@@ -112,7 +112,6 @@ void checkMatchForHighlight(TSNode node, TreePath tree_path[], int tree_path_len
   }
 
 
-
   // Get litteral string for current node. We don't extract if the litteral is higher than 200 char.
   int char_nb = ts_node_end_byte(node) - ts_node_start_byte(node);
   int array_size = char_nb;
@@ -121,12 +120,11 @@ void checkMatchForHighlight(TSNode node, TreePath tree_path[], int tree_path_len
   }
   char litteral_text_node[array_size + 1];
   if (char_nb <= REGEX_MAX_DUMP_SIZE) {
-    // strncpy(litteral_text_node, source + ts_node_start_byte(node), char_nb);
-    // fprintf(stderr, "TSPoint : start(%d, %d) => end(%d, %d)\n", ts_node_start_point(node).row, ts_node_start_point(node).column, ts_node_end_point(node).row, ts_node_end_point(node).column);
-    // fprintf(stderr, "Called from highlight\n");
     readNBytesAtPosition(tmp, ts_node_start_point(node).row, ts_node_start_point(node).column, litteral_text_node, char_nb);
   }
   litteral_text_node[array_size] = '\0';
+
+
 
   while (seq != NULL) {
     char* result = isTreePathMatchingQuery(litteral_text_node, tree_path, tree_path_length, seq->value);
